refactor: move latency_tracker to adapters/config/ and processing_mode to core/
- Move latency_tracker.py from shared/ to adapters/config/latency.py (closer to MetricsSink port)
- Move processing_mode.py from shared/ to core/processing_mode.py (domain enum belongs in core)
- Leave backward-compat shims in shared/ for any transitive importers
- Update direct importers: container.py, cli.py, session_service.py, fn/handler.py, fn/parser.py
- core/processing_mode.py reads ENABLE_ADVANCED_MODES from os.getenv to stay independent of shared.config

269 tests pass, lint clean.
